Use API to create trip

+++ app/api/common.py
@@ -1,6 +1,6 @@
 def coordinate_pair(value):
     valid_pair = 'x' in value and 'y' in value \
-                 and isinstance(value['x'], float) and isinstance(value['y'], float)
+                 and isinstance(value['x'], int) and isinstance(value['y'], int)
     if not valid_pair:
         raise ValueError("Invalid list of coordinates")
     return value
